Move component classes into their own files and use css modules
