+++ assets/js/components/Character/Gear/Header.jsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export default (props) => <div>Gear</div>
