+++ assets/js/components/Character/Gear/index.js
@@ -1,0 +1,6 @@
+import Header from './Header';
+import Tab from './Tab';
+
+export {
+    Header, Tab
+}
