+++ assets/js/components/Character/Skills/Header.jsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export default (props) => <div>Skills</div>
