+++ assets/js/components/Character/Skills/Tab.jsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export default (props) => <div>Welcome to Skills!</div>
