+++ assets/js/components/Character/Traits/Header.jsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export default (props) => <div>Traits</div>
