+++ assets/js/components/Character/Traits/Tab.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Line } from '../../Specializations';
+
+export default (props) => (<div>
+    <Line />
+    <Line />
+    <Line />
+</div>)
