+++ assets/js/components/Character/index.js
@@ -1,0 +1,11 @@
+import Panel from './Panel';
+
+import * as Gear from './Gear';
+import * as Skills from './Skills';
+import * as Traits from './Traits';
+
+export {
+    Panel,
+
+    Gear, Skills, Traits
+}
