+++ assets/js/components/Settings/Messages.jsx
@@ -1,0 +1,94 @@
+import { defineMessages } from "react-intl";
+
+export default defineMessages({
+    pve: {
+        id: "gameMode.pve",
+        description: "PvE",
+        defaultMessage: "PvE"
+    },
+    pvp: {
+        id: "gameMode.pvp",
+        description: "PvP",
+        defaultMessage: "PvP"
+    },
+    wvw: {
+        id: "gameMode.wvw",
+        description: "WvW",
+        defaultMessage: "WvW"
+    },
+    guardian: {
+        id: "profession.guardian",
+        description: "Guardian",
+        defaultMessage: "Guardian"
+    },
+    revenant: {
+        id: "profession.revenant",
+        description: "Revenant",
+        defaultMessage: "Revenant"
+    },
+    warrior: {
+        id: "profession.warrior",
+        description: "Warrior",
+        defaultMessage: "Warrior"
+    },
+    engineer: {
+        id: "profession.engineer",
+        description: "Engineer",
+        defaultMessage: "Engineer"
+    },
+    ranger: {
+        id: "profession.ranger",
+        description: "Ranger",
+        defaultMessage: "Ranger"
+    },
+    thief: {
+        id: "profession.thief",
+        description: "Thief",
+        defaultMessage: "Thief"
+    },
+    elementalist: {
+        id: "profession.elementalist",
+        description: "Elementalist",
+        defaultMessage: "Elementalist"
+    },
+    mesmer: {
+        id: "profession.mesmer",
+        description: "Mesmer",
+        defaultMessage: "Mesmer"
+    },
+    necromancer: {
+        id: "profession.necromancer",
+        description: "Necromancer",
+        defaultMessage: "Necromancer"
+    },
+    noRace: {
+        id: "race.noRace",
+        description: "No race",
+        defaultMessage: "No race"
+    },
+    asura: {
+        id: "race.asura",
+        description: "Asura",
+        defaultMessage: "Asura"
+    },
+    charr: {
+        id: "race.charr",
+        description: "Charr",
+        defaultMessage: "Charr"
+    },
+    human: {
+        id: "race.human",
+        description: "Human",
+        defaultMessage: "Human"
+    },
+    norn: {
+        id: "race.norn",
+        description: "Norn",
+        defaultMessage: "Norn"
+    },
+    sylvari: {
+        id: "race.sylvari",
+        description: "Sylvari",
+        defaultMessage: "Sylvari"
+    },
+});
