+++ assets/js/components/Settings/index.js
@@ -1,0 +1,9 @@
+import Container from './Container';
+import Header from './Header';
+import Messages from './Messages';
+import Option from './Option';
+import Panel from './Panel';
+
+export {
+    Container, Header, Messages, Option, Panel
+}
