+++ assets/js/components/Specializations/TraitLine.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import style from './trait.css';
+
+export default (props) => props.from && props.to
+    ? <div className={style[`line-${props.from}-${props.to}`]}/>
+    : <div className={style.lineInvisible}/>;
