+++ assets/js/components/Specializations/index.js
@@ -1,0 +1,9 @@
+import Line from './Line';
+import TraitGroup from './TraitGroup';
+import TraitLine from './TraitLine';
+
+export {
+    Line,
+    TraitGroup,
+    TraitLine
+}
