+++ assets/js/components/Specializations/line.css
@@ -1,0 +1,45 @@
+@value borderWidth: 2px;
+@value specHeight: 136px;
+
+.line {
+    border: borderWidth solid black;
+    display: flex;
+    align-items: center;
+    position: relative;
+    width: 647px;
+    height: specHeight;
+    margin: 2px auto;
+}
+
+.background {
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    width: 647px;
+    height: specHeight;
+}
+
+.backgroundImage {
+    transition: opacity 500ms;
+}
+
+.line:hover .backgroundImage {
+    opacity: 1;
+}
+
+.arrow {
+    display: flex;
+    align-items: center;
+    margin-right: 145px;
+    width: 16px;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.3);
+    z-index: 10;
+}
+
+.arrow > svg {
+    margin-left: 4px;
+    fill: #d8cf62;
+}
