+++ assets/js/components/Specializations/trait.css
@@ -1,0 +1,77 @@
+.block {
+    display: flex;
+    align-items: center;
+    height: 100%;
+    z-index: 10;
+}
+
+.adept, .master, .grandmaster {
+    composes: block;
+}
+
+.line {
+    width: 39px;
+    height: 2px;
+    z-index: 1;
+    background: white;
+}
+.lineInvisible {
+    composes: line;
+    background: none;
+}
+
+.line-mid-mid {
+    composes: line;
+}
+.line-mid-up {
+    composes: line;
+    transform: translateY(-20px) rotateZ(-30deg) scaleX(1.5);
+}
+.line-mid-down {
+     composes: line;
+     transform: translateY(20px) rotateZ(30deg) scaleX(1.5);
+}
+.line-up-mid {
+     composes: line;
+     transform: translateY(-20px) rotateZ(30deg) scaleX(1.5);
+}
+.line-down-mid {
+     composes: line;
+     transform: translateY(20px) rotateZ(-30deg) scaleX(1.5);
+}
+
+
+.minorIcon {
+    width: 42px;
+    height: 42px;
+    margin: 0 -4px;
+    z-index: 10;
+}
+
+.major {
+    z-index: 10;
+}
+
+
+.majorIcon {
+    background-size: 42px;
+    background-repeat: no-repeat;
+    background-image: url(https://render.guildwars2.com/file/C26CBECA553C29091F1D2C9B05732847E4C3457D/1012475.png);
+    width: 42px;
+    height: 42px;
+    margin: 0 -3px;
+    filter: brightness(0.5);
+    transition: filter 300ms;
+}
+
+.majorIcon:hover {
+    filter: brightness(0.9) sepia(10%) contrast(80%);
+    transition: filter 150ms;
+}
+
+.majorIconSelected,
+.majorIconSelected:hover {
+    composes: majorIcon;
+    filter: unset;
+    transition: filter 150ms;
+}
