+++ assets/js/components/StatsOverview/index.js
@@ -1,0 +1,5 @@
+import Panel from './Panel';
+
+export {
+    Panel
+}
