+++ assets/js/components/StatsOverview/style.css
@@ -1,0 +1,99 @@
+.panel {
+    background: lightsalmon;
+    width: 220px;
+}
+
+.list {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+    columns: 2;
+}
+
+.listitem {
+    background-size: 24px;
+    background-repeat: no-repeat;
+    padding-left: 28px;
+    margin: 2px 0;
+    height: 24px;
+    line-height: 24px;
+}
+
+/*
+Icons
+*/
+
+.attribute-power {
+    composes: listitem;
+    background-image: url(../img/stats-power.png);
+}
+
+.attribute-toughness {
+    composes: listitem;
+    background-image: url(../img/stats-toughness.png);
+}
+
+.attribute-vitality {
+    composes: listitem;
+    background-image: url(../img/stats-vitality.png);
+}
+
+.attribute-precision {
+    composes: listitem;
+    background-image: url(../img/stats-precision.png);
+}
+
+.attribute-ferocity {
+    composes: listitem;
+    background-image: url(../img/stats-ferocity.png);
+}
+
+.attribute-condition-damage {
+    composes: listitem;
+    background-image: url(../img/stats-condition-damage.png);
+}
+
+.attribute-expertise {
+    composes: listitem;
+    background-image: url(../img/stats-condition-duration.png);
+}
+
+.attribute-concentration {
+    composes: listitem;
+    background-image: url(../img/stats-boon-duration.png);
+}
+
+.attribute-armor {
+    composes: listitem;
+    background-image: url(../img/stats-armor.png);
+}
+
+.attribute-health {
+    composes: listitem;
+    background-image: url(../img/stats-health.png);
+}
+
+.attribute-critical-chance {
+    composes: listitem;
+    background-image: url(../img/stats-critical-chance.png);
+}
+
+.attribute-critical-damage {
+    composes: listitem;
+    background-image: url(../img/stats-critical-damage.png);
+}
+
+.attribute-healing-power {
+    composes: listitem;
+    background-image: url(../img/stats-healing-power.png);
+}
+
+.attribute-condition-duration {
+    composes: listitem;
+    background-image: url(../img/stats-condition-duration.png);
+}
+
+.attribute-boon-duration {
+    composes: listitem;
+    background-image: url(../img/stats-boon-duration.png);
+}
