--- assets/sass/app.scss
@@ -1,25 +1,0 @@
-@import "components/index";
-
-* {
-  box-sizing: border-box;
-}
-
-body {
-  margin: 0;
-  padding: 0;
-  font-family: Verdana, Tahoma, Geneva, Arial, sans-serif;
-  font-size: 10pt;
-}
-
-.editor {
-  border: 1px solid black; // TODO: Remove later
-  position: relative;
-  width: 960px;
-  height: 520px;
-  margin: auto;
-  cursor: default;
-}
-
-.hidden {
-  visibility: hidden;
-}
