--- assets/sass/components/base/index.scss
@@ -1,2 +1,0 @@
-@import "tabs-container";
-@import "overlay-container";
