--- assets/sass/components/index.scss
@@ -1,2 +1,0 @@
-@import "base/index";
-@import "panels/index";
