--- assets/sass/components/panels/character.scss
@@ -1,22 +1,0 @@
-.character-panel {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-
-  .tab-container-headers {
-    background: lightgreen; // TODO: Remove later
-    height: 40px;
-    margin: 0 220px;
-  }
-
-  .tab-container-active-tab {
-    background: lightpink; // TODO: Remove later
-    position: absolute;
-    top: 40px;
-    bottom: 0;
-    left: 0;
-    right: 220px;
-  }
-}
