--- assets/sass/components/panels/index.scss
@@ -1,5 +1,0 @@
-@import "settings";
-@import "character";
-@import "traits";
-@import "stats-overview";
-@import "professions/index";
