--- assets/sass/components/panels/professions/index.scss
@@ -1,9 +1,0 @@
-@import "guardian";
-//@import "revenant";
-//@import "warrior";
-//@import "engineer";
-//@import "ranger";
-//@import "thief";
-//@import "ementalist";
-//@import "mesmer";
-//@import "necromancer";
