--- assets/sass/components/panels/traits.scss
@@ -1,95 +1,0 @@
-.specializations {
-  .specialization-line {
-    $borderWidth: 1px;
-    $specHeight: 136px;
-    border: $borderWidth solid black; // TODO: Remove later
-    display: flex;
-    align-items: center;
-    position: relative;
-    width: 647px; // Magic calculation, don't touch
-    height: $specHeight; // Another magic calculation
-    margin: 2px auto;
-
-    .specialization-background {
-      position: absolute;
-      top: 0;
-      bottom: 0;
-      left: 0;
-      right: 0;
-      width: 647px;
-      height: $specHeight;
-      #background {
-        transition: opacity 500ms;
-      }
-    }
-    &:hover .specialization-background #background {
-      opacity: 1;
-    }
-    .specialization-arrow {
-      display: flex;
-      align-items: center;
-      margin-right: 145px;
-      width: 16px;
-      height: 100%;
-      background: rgba(0, 0, 0, 0.3);
-      z-index: 10;
-      .specialization-arrow-icon {
-        margin-left: 4px;
-        fill: #d8cf62;
-      }
-    }
-    .specialization-traits-block {
-      display: flex;
-      align-items: center;
-      height: 100%;
-      z-index: 10;
-      .specialization-traits-line {
-        width: 39px;
-        height: 2px;
-        z-index: 1;
-        &.mid-up, &.mid-down, &.mid-mid, &.up-mid, &.down-mid {
-          background: white;
-        }
-        &.mid-up {
-          transform: translateY(-20px) rotateZ(-30deg) scaleX(1.5);
-        }
-        &.mid-down {
-          transform: translateY(20px) rotateZ(30deg) scaleX(1.5);
-        }
-        &.up-mid {
-          transform: translateY(-20px) rotateZ(30deg) scaleX(1.5);
-        }
-        &.down-mid {
-          transform: translateY(20px) rotateZ(-30deg) scaleX(1.5);
-        }
-      }
-      .specialization-traits-minor-icon {
-        width: 42px;
-        height: 42px;
-        margin: 0 -4px;
-        z-index: 10;
-      }
-      .specialization-traits-major {
-        z-index: 10;
-      }
-      .specialization-traits-major-icon {
-        background-size: 42px;
-        background-repeat: no-repeat;
-        background-image: url(https://render.guildwars2.com/file/C26CBECA553C29091F1D2C9B05732847E4C3457D/1012475.png); // TODO: Remove later
-        width: 42px;
-        height: 42px;
-        margin: 0 -3px;
-        filter: brightness(0.5);
-        transition: filter 300ms;
-        &:hover {
-          filter: brightness(0.9) sepia(10%) contrast(80%);
-          transition: filter 150ms;
-        }
-        &.selected, &.selected:hover {
-          filter: unset;
-          transition: filter 150ms;
-        }
-      }
-    }
-  }
-}
