+++ package.json
@@ -21,6 +21,7 @@
     "serve-favicon": "^2.3.0"
   },
   "devDependencies": {
+    "autoprefixer": "^6.5.1",
     "babel-plugin-lodash": "^3.2.9",
     "babel-plugin-react-intl": "^2.2.0",
     "babel-preset-es2015": "^6.16.0",
@@ -29,19 +30,19 @@
     "browser-sync": "^2.17.5",
     "browserify": "^13.1.0",
     "change-case": "^3.0.0",
+    "cssnano": "^3.8.0",
     "del": "^2.2.2",
     "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^3.1.1",
     "gulp-babel": "^6.1.2",
     "gulp-imagemin": "^3.0.3",
     "gulp-nodemon": "^2.2.1",
     "gulp-rev-all": "^0.9.7",
-    "gulp-sass": "^2.3.2",
     "gulp-sourcemaps": "^2.1.1",
     "gulp-uglify": "^2.0.0",
     "gulp-util": "^3.0.7",
     "gw2api-client": "^3.1.2",
     "lodash": "^4.16.4",
+    "modular-css": "^0.28.2",
     "react": "^15.3.2",
     "react-addons-css-transition-group": "^15.3.2",
     "react-addons-transition-group": "^15.3.2",
